requirements.txt
streamlit==1.26.0
streamlit-chat==0.1.1
streamlit-extras==0.3.0
llama-cpp-python==0.1.83